[refactor] Split code into modules
Is this better? Who knows. But let's try it.
